+++ project/test/unit/10CreationSocietesTest.php
@@ -14,7 +14,9 @@ foreach (CompteTagsView::getInstance()->listByTags('test', 'test') as $k => $v) 
         //     $drm = DRMClient::getInstance()->find($id);
         //     $drm->delete(false);
         //   }
-        $etabl->etablissement->delete();
+        if ($etabl->etablissement) {
+            $etabl->etablissement->delete();
+        }
       }
       $soc->delete();
     }
